gmapping/requirements.txt
actionlib==1.14.0
Adafruit-Blinka==8.39.1
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-connectionmanager==1.0.1
adafruit-circuitpython-icm20x==2.0.20
adafruit-circuitpython-register==1.9.18
adafruit-circuitpython-requests==3.2.5
adafruit-circuitpython-typing==1.10.3
Adafruit-PlatformDetect==3.62.0
Adafruit-PureIO==1.1.11
AHRS==0.3.1
angles==1.9.13
apt-clone==0.2.1
apturl==0.5.2
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
bcrypt==3.1.7
blinker==1.7.0
bondpy==1.8.6
Brlapi==0.7.0
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
cbor==1.0.0
certifi==2019.11.28
chardet==3.0.4
click==8.1.7
colorama==0.4.3
constantly==15.1.0
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cryptography==2.8
cupshelpers==1.0
cv-bridge==1.16.2
cycler==0.10.0
Cython==0.29.14
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distro==1.4.0
distro-info==0.23+ubuntu1.1
docutils==0.16
duplicity==0.8.12.0
dynamic-reconfigure==1.7.3
dynamixel-sdk==3.7.31
easyocr==1.7.1
empy==3.3.2
entrypoints==0.3
fasteners==0.14.1
filelock==3.13.4
flask==3.0.3
fsspec==2024.3.1
future==0.18.2
gazebo-plugins==2.9.2
gazebo-ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
graphsurgeon==0.4.6
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
image-geometry==1.16.2
imageio==2.34.0
importlib-metadata==7.1.0
incremental==16.10.1
interactive-markers==1.12.0
itsdangerous==2.2.0
Jetson.GPIO==2.1.6
Jinja2==3.1.3
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
laser-geometry==1.6.7
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-loader==0.4
lockfile==0.12.2
louis==3.12.0
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.5
matplotlib==3.1.2
message-filters==1.16.0
monotonic==1.5
mpi4py==3.0.3
mpmath==1.3.0
netifaces==0.10.4
networkx==3.1
ninja==1.11.1.1
nose==1.3.7
numpy==1.21.1
oauthlib==3.1.0
olefile==0.46
onboard==1.4.1
onnx==1.16.0
onnx-graphsurgeon==0.3.12
opencv-python-headless==4.9.0.80
packaging==24.0
PAM==0.4.2
pandas==0.25.3
paramiko==2.6.0
pexpect==4.6.0
Pillow==9.1.0
protobuf==5.26.1
psutil==5.5.1
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pyclipper==1.3.0.post5
pycrypto==2.6.1
pycryptodomex==3.6.1
pycups==1.9.73
pydot==1.4.1
pyftdi==0.55.4
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyICU==2.4.2
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.14.1
pyRFC3339==1.1
pyserial==3.5
python-apt==2.0.1+ubuntu0.20.4.1
python-bidi==0.4.2
python-dateutil==2.7.3
python-dbusmock==0.19
python-debian==0.1.36+ubuntu1.1
python-gnupg==0.4.5
python-qt-binding==0.4.4
python-snappy==0.5.3
PyTrie==0.2
pytz==2019.3
pyusb==1.2.1
PyWavelets==1.4.1
pyxdg==0.26
PyYAML==5.3.1
pyzbar==0.1.9
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.22.0
requests-unixsocket==0.2.0
resource-retriever==1.12.7
roman==2.0.0
rosbag==1.16.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosdep==0.22.2
rosdep-modules==0.22.2
rosdistro-modules==0.9.0
rosgraph==1.16.0
roslaunch==1.16.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.16.0
rosmake==1.15.8
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospkg==1.5.0
rospkg-modules==1.5.0
rospy==1.16.0
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.12
rqt-dep==0.4.12
rqt-graph==0.4.14
rqt-gui==0.5.3
rqt-gui-py==0.5.3
rqt-image-view==0.4.17
rqt-launch==0.4.9
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-msg==0.4.10
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.11
rqt-publisher==0.4.10
rqt-py-common==0.5.3
rqt-py-console==0.4.10
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-robot-steering==0.5.12
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.0
rqt-service-caller==0.4.10
rqt-shell==0.4.11
rqt-srv==0.4.9
rqt-tf-tree==0.6.4
rqt-top==0.4.10
rqt-topic==0.4.13
rqt-web==0.4.10
rviz==1.14.20
scikit-image==0.21.0
scipy==1.8.0
SecretStorage==2.3.1
sensor-msgs==1.13.1
service-identity==18.1.0
shapely==2.0.4
simplejson==3.16.0
sip==4.19.21
six==1.14.0
smach==2.5.2
smach-ros==2.5.2
smclib==1.8.6
sympy==1.12
systemd-python==234
tensorrt==8.5.2.2
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
tifffile==2023.7.10
topic-tools==1.16.0
torch==2.2.2
torchvision==0.17.2
Twisted==18.9.0
txaio==2.10.0
typing-extensions==4.11.0
u-msgpack-python==2.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
uff==0.6.9
urllib3==1.25.8
urwid==2.0.1
wadllib==1.3.3
werkzeug==3.0.2
wsaccel==0.6.2
xacro==1.14.17
xkit==0.0.0
zipp==3.18.1
zope.interface==4.7.1
